+++ R/get_varcov_sandwich.R
@@ -78,7 +78,9 @@
 
   ## TODO: what about Sattherthwaite? @vincentarelbundock
 
-  if (!grepl("^(vcov|kernHAC|NeweyWest)", vcov_fun)) {
+  if (grepl("^(vcov|kernHAC|NeweyWest)", vcov_fun)) {
+    vcov_fun_clean <- vcov_fun
+  } else {
     vcov_fun_clean <- switch(vcov_fun,
       HC0 = ,
       HC1 = ,
@@ -108,8 +110,6 @@
       PL = "vcovPL",
       `kenward-roger` = "vcovAdj"
     )
-  } else {
-    vcov_fun_clean <- vcov_fun
   }
 
   # check if required package is available
@@ -139,9 +139,9 @@
 
   # extract variance-covariance matrix
   if (!inherits(.vcov, "matrix")) {
-    msg <- sprintf("Unable to extract a variance-covariance matrix for model object of class `%s`. Different values of the `vcov` argument trigger calls to the `sandwich` or `clubSandwich` packages in order to extract the matrix (see `?insight::get_varcov`). Your model or the requested estimation type may not be supported by one or both of those packages, or you were missing one or more required arguments in `vcov_args` (like `cluster`).", class(x)[1])
+    msg <- sprintf("Unable to extract a variance-covariance matrix for model object of class `%s`. Different values of the `vcov` argument trigger calls to the `sandwich` or `clubSandwich` packages in order to extract the matrix (see `?insight::get_varcov`). Your model or the requested estimation type may not be supported by one or both of those packages, or you were missing one or more required arguments in `vcov_args` (like `cluster`).", class(x)[1]) # nolint
     if (inherits(.vcov, "try-error")) {
-      msg <- c(msg, "", "This error was raised:", "", attr(.vcov, "condition")$message)
+      msg <- c(msg, "", "This error was raised:", attr(.vcov, "condition")$message)
     }
     format_error(msg)
   }